Small return value adjustments

+++ src/tacho.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void tacho_init();
-uint16_t tacho_get_speed();
+uint8_t tacho_get_speed();
 uint16_t tacho_get_distance();
 void tacho_reset_distance();
